move common get_resource_scheduled_date to Delivery Utils module
