+++ lib/oli_web/components/delivery/course_lastest_visited_page.ex
@@ -1,8 +1,8 @@
 defmodule OliWeb.Components.Delivery.CourseLatestVisitedPage do
   use OliWeb, :html
 
-  alias OliWeb.Router.Helpers, as: Routes
   alias OliWeb.Components.Delivery.Buttons
+  alias OliWeb.Components.Delivery.Utils, as: DeliveryUtils
 
   attr :page_revision, Oli.Resources.Revision
   attr :section_slug, :string
@@ -25,7 +25,7 @@ defmodule OliWeb.Components.Delivery.CourseLatestVisitedPage do
 
           <%= if !@is_instructor do %>
             <span class="w-64 h-10 text-sm tracking-wide text-gray-800 dark:text-white bg-gray-100 dark:bg-gray-500 rounded-sm flex justify-center items-center ml-auto mr-3">
-              <%= get_resource_scheduled_date(@page_revision.resource_id, @scheduled_dates, @ctx) %>
+              <%= DeliveryUtils.get_resource_scheduled_date(@page_revision.resource_id, @scheduled_dates, @ctx) %>
             </span>
             <.link navigate={~p"/sections/#{@section_slug}/page/#{@page_revision.slug}"} class="torus-button primary h-10">Open</.link>
           <% else %>
@@ -49,19 +49,4 @@ defmodule OliWeb.Components.Delivery.CourseLatestVisitedPage do
     </div>
     """
   end
-
-  # TODO move this and the same functions in course_content.ex to utils.ex of this folder
-  defp get_resource_scheduled_date(resource_id, scheduled_dates, ctx) do
-    case scheduled_dates[resource_id] do
-      %{end_date: nil} ->
-        "No due date"
-
-      data ->
-        "#{scheduled_date_type(data.scheduled_type)} #{OliWeb.Common.FormatDateTime.date(data.end_date, ctx)}"
-    end
-  end
-
-  defp scheduled_date_type(:read_by), do: "Read by"
-  defp scheduled_date_type(:inclass_activity), do: "In class on"
-  defp scheduled_date_type(_), do: "Due by"
 end
